refactor: improve balance precision and optimize asset lookups
- Replace Double with BigDecimal for fiat values to prevent rounding errors.
- Refactor EVM and SPL balances to use Maps for O(1) lookup performance.
- Centralize crypto, currency, and timestamp formatting into FormattingUtils.kt.
- Delegate TransactionFormatHelper logic to centralized formatting extensions.

+++ app/src/main/java/com/example/nexuswallet/feature/wallet/domain/model/AssetDisplayInfo.kt
@@ -1,5 +1,7 @@
 package com.example.nexuswallet.feature.wallet.domain.model
 
+import java.math.BigDecimal
+
 data class AssetDisplayInfo(
     val id: String,
     val walletId: String,
@@ -10,7 +12,7 @@ data class AssetDisplayInfo(
     val isTestnet: Boolean,
     val balance: String,
     val balanceFormatted: String,
-    val usdValue: Double,
+    val usdValue: BigDecimal,
     val usdValueFormatted: String,
     val priceChangePercentage: Double?,
     val priceChangeFormatted: String?,
